pisah gui dan main
